Support for docker
This commit adds minimal support for docker. The goal is to have an
automated build in Docker hub that is automatically updated when an
change is made to the bitcoinz code base. Developers can then use
this to extend and write apps around the bitcoinz app using docker.

Cli command

This commit creates a cli wrapper script so that the zcash-cli tool
tha can be executed against a running docker container.