Add support for Snowgem's ModernWallet
Summary:
In order for Snowgem to be able to add BitcoinZ to their ModernWallet,
they need us to add a getalldata command that provides it with all the
information it needs quickly. This command enables this.

Test Plan:
I've compiled this in a node node and verified that I was able to see
the new command in help, as well as execute it to get the expected
response.  A sample of the response data looks as follows:

{
  "connectionCount": 3,
  "besttime": 1513645152,
  "bestblockhash": "0000000339ab2556d7a8af4c83b9c4d6771e6efc4b1c284b2b6705f61339736b",
  "transparentbalance": "0.00",
  "privatebalance": "0.00",
  "lockedbalance": "0.00",
  "totalbalance": "0.00",
  "unconfirmedbalance": "0.00",
  "immaturebalance": "0.00",
  "addressbalance": [
    {
      "": 0.00000000
    }
  ],
  "listtransactions": [
    {
      "account": "",
      "address": "",
      "category": "",
      "amount": "0",
      "vout": "1",
      "confirmations": "0",
      "generated": "true",
      "blockhash": "0000000000000000000000000000000000000000000000000000000000000000",
      "blockindex": "0",
      "blocktime": "0",
      "txid": "0000000000000000000000000000000000000000000000000000000000000000",
      "time": "0",
      "timereceived": "0"
    }
  ]
}