Can't divide by zero
When mining blocks using the tromp solver, occasionally the solver
returns a 0 for Estimate Net Height. Since the metrics use the
real height / Estimated Net Height, this causes problems when the
value is 0 because maths. This commit adds a check to make sure the
value of EstimateNetHeight is not zero before allowing division.